requirements.txt
# Agro AI - Python Dependencies
# Install with: pip install -r requirements.txt

flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
